[emapp] make nanoem_pragma_diagnostics* public  (#253)
* [emapp] make `nanoem_pragma_diagnostics*` public

The commit also reduces several warnings caused by internals and
external libraries.

* [emapp] fix build failure on win32

* [emapp] replaces `emapp/private/imgui/config.h` with its content

* [emapp] implements missing timestamp method on Linux

* [emapp] reduce warnings on MSVC

* [emapp] fix build on win32

+++ emapp/test/misc/stringutils.cc
@@ -81,7 +81,7 @@ TEST_CASE("stringutils_formatoverflow", "[emapp][misc]")
     for (int i = 0; i <= 16384; i++) {
         sb.append("0");
     }
-    CHECK(StringUtils::format(result, "%s", sb.c_str()) == sb.size());
+    CHECK(StringUtils::format(result, "%s", sb.c_str()) == static_cast<int>(sb.size()));
     CHECK(result == sb);
 }
